requirements.txt
Django==2.2.13
djangorestframework
Pillow
celery==3.1.25
redis
requests
django-celery-beat==1.1.0
pyyaml
uritemplate